Moved hysteresis methods.
