+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Changes on the `main` branch, but not yet released, will be listed here.
 
 -   [[#27](https://github.com/diatche/LibreChart/pull/27)] When specifying an axes on a plot, passing `true` for an axis-type will add a default axis, and passing `false` or `undefined` for an axis-type will not add an axis.
 
+### Breaking Changes
+
+-   The type `ScaleHysteresisFunction` is now `Hysteresis.StepFunc`.
+
 ## 0.5.0
 
 **17 Mar 2021**
